chore(examples): JSE-18 refactor sandbox config logic; group into a service and split by framework
